Struct cryptatools_core::cryptanalysis::general_cryptanalysis_methods::frequency_analysis::distribution_algorithms::statistical::Statistical

pub struct Statistical { /* private fields */ }

Implementations

impl Statistical

pub fn new(alphabet: HashMap<String, Vec<u8>>) -> Self

pub fn guess_statistical_distribution(
&self,
encrypted_input: Vec<u8>
) -> HashMap<Vec<u8>, f64>

Catch statistical distribution (percentage) from a encrypted text.

Simply count the occurences of a byte in a cipher text. Then divide the result by the cipher text size.

use std::collections::HashMap;
use cryptatools_core::cryptanalysis::general_cryptanalysis_methods::frequency_analysis::distribution_algorithms::statistical::{Statistical};
use once_cell::sync::Lazy;
use cryptatools_core::utils::alphabets::ASCII_ALPHABET;
let stat = Statistical::new(Lazy::force(&ASCII_ALPHABET).to_owned());
let stat_percentage = stat.guess_statistical_distribution(vec![0x41, 0x41, 0x41, 0x41, 0x41, 0x41]);
assert_eq!(stat_percentage, HashMap::from([(vec![0x41 as u8], 1.0)]));

use std::collections::HashMap;
use cryptatools_core::cryptanalysis::general_cryptanalysis_methods::frequency_analysis::distribution_algorithms::statistical::{Statistical};
use once_cell::sync::Lazy;
use cryptatools_core::utils::alphabets::ASCII_ALPHABET;
let stat = Statistical::new(Lazy::force(&ASCII_ALPHABET).to_owned());
let stat_percentage = stat.guess_statistical_distribution(vec![0x41, 0x41, 0x41, 0x41, 0x41, 0x42]);
assert_eq!(stat_percentage, HashMap::from([(vec![0x42 as u8], 0.16666666666666666), (vec![0x41 as u8], 0.8333333333333334)]));