cryptatools_core::cryptography::classical::encryption::monoalphabetic_ciphers::caesar_number

Struct CaesarNumberAlgorithm

pub struct CaesarNumberAlgorithm {
    pub alphabet: Arc<Alphabet>,
}

Fields
alphabet: Arc<Alphabet>

Alphabet used by the caesar number encryption Algotithm.

Implementations

impl CaesarNumberAlgorithm

pub fn new(alphabet: Arc<Alphabet>) -> Self

pub fn encrypt_by_opcode_shift(&self, plain_text: Vec<u8>, key: u32) -> Vec<u8>

Encrypt the plain text with the caesar number encryption algorithm. Does not care of the alphabet.

The plain_text is passed as argument. Each character in the plain_text is shifted of key ranges in his opcode representation. If the alphabet overflows, then the cipher text continues from the start of the alphabet. The custom alphabet has been put in the constructor of the struct CaesarNumberAlgorithm.

use cryptatools_core::cryptography::classical::encryption::monoalphabetic_ciphers::caesar_number::CaesarNumberAlgorithm;
use cryptatools_core::utils::alphabets::Alphabet;
  
let ascii_alphabet = Alphabet::new_empty().ascii_encoding();
let mut c: CaesarNumberAlgorithm = CaesarNumberAlgorithm::new(ascii_alphabet.into());
let encrypted = c.encrypt_by_opcode_shift(vec![0x41, 0x41, 0x41], 1);
assert_eq!(vec![0x42, 0x42, 0x42], encrypted);

use cryptatools_core::cryptography::classical::encryption::monoalphabetic_ciphers::caesar_number::CaesarNumberAlgorithm;
use cryptatools_core::utils::alphabets::Alphabet;
use std::char;
 
let ascii_alphabet = Alphabet::new_empty().ascii_encoding();
let mut c: CaesarNumberAlgorithm = CaesarNumberAlgorithm::new(ascii_alphabet.into());
let plain_text: Vec<u8> = vec![0x41, 0x41, 0x41];
let encrypted = c.encrypt_by_opcode_shift(plain_text, 1);
let mut re_encrypted = String::new();
for character_int in encrypted {
    re_encrypted.push(character_int.into());
}
assert_eq!(re_encrypted, "BBB");

use cryptatools_core::cryptography::classical::encryption::monoalphabetic_ciphers::caesar_number::CaesarNumberAlgorithm;
use cryptatools_core::utils::alphabets::Alphabet;
use std::char;
 
let printable_chars_of_ascii_characters = Alphabet::new_empty().ascii_printable_only_encoding();
let mut c: CaesarNumberAlgorithm = CaesarNumberAlgorithm::new(printable_chars_of_ascii_characters.into());
let plain_text: Vec<u8> = vec![0x41, 0x41, 0x41];
let encrypted = c.encrypt_by_opcode_shift(plain_text, 10);
let mut re_encrypted = String::new();
for character_int in encrypted {
    re_encrypted.push(character_int.into());
}
assert_eq!(re_encrypted, "KKK");

pub fn decrypt_by_opcode_shift(&self, cipher_text: Vec<u8>, key: u32) -> Vec<u8>

Decrypt the cipher text with the caesar number encryption algorithm. Does not care of the alphabet.

The cipher_text is passed as argument. Each character in the cipher_text is itself minus key. If the alphabet overflows, then the cipher text continues from the end of the alphabet. The custom alphabet has been put in the constructor of the struct CaesarNumberAlgorithm.

use cryptatools_core::cryptography::classical::encryption::monoalphabetic_ciphers::caesar_number::CaesarNumberAlgorithm;
use cryptatools_core::utils::alphabets::Alphabet;
  
let ascii_alphabet = Alphabet::new_empty().ascii_encoding();
let mut c: CaesarNumberAlgorithm = CaesarNumberAlgorithm::new(ascii_alphabet.into());
let encrypted = c.encrypt_by_opcode_shift(vec![0x41, 0x41, 0x41], 1);
assert_eq!(vec![0x42, 0x42, 0x42], encrypted);

pub fn encrypt_by_alphabet_shift( &self, plain_text: Vec<u8>, key: u32, ) -> Vec<u8>

Encrypt the plain text with the caesar number encryption algorithm.

The plain_text is passed as argument. Each character in the plain_text is shifted of key ranges in the alphabet. If the alphabet overflows, then the cipher text continues from the start of the alphabet. The custom alphabet has been put in the constructor of the struct CaesarNumberAlgorithm.