Struct cryptatools_core::cryptography::classical::encryption::transpositional_ciphers::columnar_transposition::ColumnarTranspositionAlgorithm

pub struct ColumnarTranspositionAlgorithm { /* private fields */ }

Implementations

impl ColumnarTranspositionAlgorithm

pub fn new(alphabet: Arc<Alphabet>) -> Self

pub fn encrypt(&self, plain_text: Vec<u8>, key: u32) -> Vec<u8>

Encrypt the plain_text with the columnar transposition encryption algorithm key.

An array with lines of size key is created with the plain_text argument. The cipher text is producted as output with reading the plain text, column by column, from top to bot, from lefgt to right. The custom alphabet has been put in the constructor of the struct CaesarNumberAlgorithm.

use cryptatools_core::utils::alphabets::Alphabet;
use cryptatools_core::utils::convert::{Encode, Decode};
use cryptatools_core::cryptography::classical::encryption::transpositional_ciphers::columnar_transposition::ColumnarTranspositionAlgorithm;
 
let printable_ascii_alphabet = Alphabet::new_empty().ascii_printable_only_encoding();
let key = 6;
let plain_text = "all work and no play makes johnny a dull boy".replace(" ", "").to_uppercase();
let plain_text_opcodes = Encode::encode(&printable_ascii_alphabet, String::from(plain_text));
let mut columnar = ColumnarTranspositionAlgorithm::new(printable_ascii_alphabet.into());
let cipher_text_opcodes = columnar.encrypt(plain_text_opcodes, 6);
let ascii_plain_text = Decode::from_u8_to_ascii(cipher_text_opcodes);
 
assert_eq!(ascii_plain_text, "AKPKNLLALENLLNASYBWDYJAOONMODYROAHU");

pub fn decrypt(&self, cipher_text: Vec<u8>, key: u32) -> Vec<u8>

Decrypt the cipher_text with the columnar transposition encryption algorithm key.

Exatcly same operation as encrypt method.

use cryptatools_core::utils::alphabets::Alphabet;
use cryptatools_core::utils::convert::{Encode, Decode};
use cryptatools_core::cryptography::classical::encryption::transpositional_ciphers::columnar_transposition::ColumnarTranspositionAlgorithm;

let printable_ascii_alphabet = Alphabet::new_empty().ascii_printable_only_encoding();
let mut columnar = ColumnarTranspositionAlgorithm::new(printable_ascii_alphabet.into());
let cipher_text_ascii = String::from("AKPKNLLALENLLNASYBWDYJAOONMODYROAHU");
let opcodes_plain_text = Encode::from_ascii_to_u8(cipher_text_ascii);
let key = 6;
let plain_text_opcodes = columnar.decrypt(opcodes_plain_text, key);
let ascii_plain_text = Decode::from_u8_to_ascii(plain_text_opcodes);
 
assert_eq!(ascii_plain_text, "all work and no play makes johnny a dull boy".replace(" ", "").to_uppercase());