cryptatools_core::utils::alphabets

Function split_bytes_by_characters_representation

pub fn split_bytes_by_characters_representation(
    encoding: &Alphabet,
    text: Vec<u8>,
) -> Vec<Vec<u8>>

 use cryptatools_core::utils::{convert::Encode, alphabets::split_bytes_by_characters_representation, alphabets::Alphabet};
 
 let ascii_alphabet = Alphabet::new_empty().ascii_encoding();
 let plaintext = split_bytes_by_characters_representation(&ascii_alphabet, Encode::from_ascii_to_u8(String::from("the quick brown roman fox jumped over the lazy ostrogoth dog")));
 assert_eq!(plaintext, [[116], [104], [101], [32], [113], [117], [105], [99], [107], [32], [98], [114], [111], [119], [110], [32], [114], [111], [109], [97], [110], [32], [102], [111], [120], [32], [106], [117], [109], [112], [101], [100], [32], [111], [118], [101], [114], [32], [116], [104], [101], [32], [108], [97], [122], [121], [32], [111], [115], [116], [114], [111], [103], [111], [116], [104], [32], [100], [111], [103]]);
 assert_eq!(plaintext.len(), String::from("the quick brown roman fox jumped over the lazy ostrogoth dog").len());